requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
scikit-learn
SQLAlchemy==2.0.7
python-dotenv==1.0.0
psycopg2==2.9.5